users can only delete themselves

+++ src/main/java/com/scottlogic/training/user/UserController.java
@@ -1,6 +1,9 @@
 package com.scottlogic.training.user;
 
+import com.scottlogic.training.auth.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -9,6 +12,9 @@ import java.util.List;
 public class UserController {
     @Autowired
     UserService userService;
+
+    @Autowired
+    AuthService authService;
 
     @GetMapping("/generateUsers")
     private List<User> generateUsers(@RequestParam("number") int numberOfUsers) {
@@ -26,8 +32,15 @@ public class UserController {
     }
 
     @DeleteMapping("/user/{username}")
-    private void deleteUser(@PathVariable("username") String username) {
+    private ResponseEntity<String> deleteUser(
+            @RequestHeader(value = "Authorization") String authorisation,
+            @PathVariable("username") String username) {
+        String authUsername = authService.getUsername(authorisation);
+        if (!authUsername.equals(username)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         userService.delete(username);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/user")
